Bug fix & new functionality
Insert had a bug where the balance of the child was not being checked for the left rotate case, it is fixed. Insertion now also outputs the success, which denotes if the key to be added was already in the tree, which is used by the insertion interface function to increase the length of the tree. Report debug function now also outputs the status of the tree, i.e. whether it is sorted or not and whether it is balanced or not.