Renamed main function in AVL Tree
Test bench main function is renamed as avlTest to be included in some master code with extern. Reason for this change is that a project may need to include more than one data structure test benches, or simply calls the test bench somewhere but does other things as well, would have its own main function which will contradict with the test bench's main.